Med-9 Refactorings.
