+++ app/hooks/useCache.ts
@@ -1,18 +1,17 @@
 import {MutableRefObject, useEffect, useRef} from "react";
-import {CacheEntry} from "@/app/types/Payout";
+import {CacheEntry} from "@/app/types";
 
+/**
+ * implement useCache hook to cache a value
+ */
 const DEFAULT_INVALIDATION_TIMEOUT = 20000; // 20 seconds
 
-export default function useCache<T>(invalidationTimeout : number = DEFAULT_INVALIDATION_TIMEOUT) {
+export default function useCache<T>(invalidationTimeout: number = DEFAULT_INVALIDATION_TIMEOUT) {
     const cache = useRef(new Map<string, CacheEntry<T>>());
-    const invalidateCacheTimeout : MutableRefObject<number | undefined> = useRef(undefined);
+    const invalidateCacheTimeout: MutableRefObject<number | undefined> = useRef(undefined);
     const clearInvalidatedCacheEntries = () => {
-        console.log(">> cache.current", cache.current)
-        console.log(">> clearInvalidatedCacheEntries")
         const now = Date.now()
-        // @ts-ignore
         for (const [key, cacheEntry] of cache.current.entries()) {
-            console.log(">> diff", now - cacheEntry.cachedAt.getTime());
             if (now - cacheEntry.cachedAt.getTime() >= invalidationTimeout) {
                 cache.current.delete(key);
             }
@@ -24,7 +23,7 @@ export default function useCache<T>(invalidationTimeout : number = DEFAULT_INVAL
         return () => clearTimeout(invalidateCacheTimeout.current)
     }, []);
 
-    return async(key: string, loader: () => Promise<T>) => {
+    return async (key: string, loader: () => Promise<T>) => {
         if (cache.current.has(key)) {
             return cache.current.get(key)?.entry as T;
         }
@@ -35,8 +34,6 @@ export default function useCache<T>(invalidationTimeout : number = DEFAULT_INVAL
             entry,
         });
 
-        console.log(">> cache.current", cache.current)
-        console.log(">> settimeout")
         clearTimeout(invalidateCacheTimeout.current);
         invalidateCacheTimeout.current = window.setTimeout(clearInvalidatedCacheEntries, invalidationTimeout + 1000);
