+++ app/hooks/useDebounce.ts
@@ -2,19 +2,19 @@
  * implement useDebounce hook to debounce a value
  */
 
-import { useEffect, useState } from "react";
+import {useEffect, useState} from "react";
 
 export const useDebounce = (value: string, delay: number) => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
+    const [debouncedValue, setDebouncedValue] = useState(value);
 
-  useEffect(() => {
-    const timeoutId = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [value, delay]);
+    useEffect(() => {
+        const timeoutId = setTimeout(() => {
+            setDebouncedValue(value);
+        }, delay);
+        return () => {
+            clearTimeout(timeoutId);
+        };
+    }, [value, delay]);
 
-  return debouncedValue;
+    return debouncedValue;
 };
